+++ src/components/sbb-link/sbb-link.stories.js
@@ -8,9 +8,9 @@ import {
   SbbTypoScaleDefault,
   SbbTypoTypeFaceSbbRoman,
 } from '@sbb-esta/lyne-design-tokens';
-import getMarkupForSvg from '../../global/helpers/get-markup-for-svg';
 import { h } from 'jsx-dom';
 import readme from './readme.md';
+import events from './sbb-link.events';
 
 const wrapperStyle = (context) => {
   if (!context.args.negative) {
@@ -20,12 +20,7 @@ const wrapperStyle = (context) => {
   return `background-color: ${SbbColorCharcoalDefault};`;
 };
 
-const Template = (args) => (
-  <sbb-link {...args}>
-    {args.icon && <span slot="icon">{getMarkupForSvg(args.icon)}</span>}
-    {args.text}
-  </sbb-link>
-);
+const Template = (args) => <sbb-link {...args}>{args.text}</sbb-link>;
 
 const paragraphStyle = (context) => {
   let color;
@@ -70,15 +65,6 @@ const download = {
 const disabled = {
   control: {
     type: 'boolean',
-  },
-};
-
-const iconFlip = {
-  control: {
-    type: 'boolean',
-  },
-  table: {
-    category: 'Icon',
   },
 };
 
@@ -132,13 +118,24 @@ const textSize = {
   options: ['xs', 's', 'm'],
 };
 
+const type = {
+  control: {
+    type: 'select',
+  },
+  options: ['button', 'reset', 'submit'],
+};
+
+const eventId = {
+  control: {
+    type: 'text',
+  },
+};
+
 const defaultArgTypes = {
   download,
-  disabled,
   negative,
   href: href,
   icon,
-  'icon-flip': iconFlip,
   'icon-placement': iconPlacement,
   'id-value': idValue,
   'accessibility-label': accessibilityLabel,
@@ -153,7 +150,6 @@ const defaultArgs = {
   negative: false,
   href: 'https://github.com/lyne-design-system/lyne-components',
   icon: '',
-  'icon-flip': false,
   'icon-placement': iconPlacement.options[0],
   'id-value': '',
   'accessibility-label': 'Travelcards & tickets',
@@ -220,7 +216,6 @@ BlockIconStart.argTypes = defaultArgTypes;
 BlockIconStart.args = {
   ...defaultArgs,
   icon: 'chevron-small-left-small',
-  'icon-flip': true,
   'text-size': textSize.options[0],
 };
 
@@ -234,7 +229,6 @@ BlockNegativeIconStart.argTypes = defaultArgTypes;
 BlockNegativeIconStart.args = {
   ...defaultArgs,
   icon: 'chevron-small-left-small',
-  'icon-flip': true,
   'text-size': textSize.options[0],
   negative: true,
 };
@@ -249,7 +243,6 @@ BlockIconEnd.argTypes = defaultArgTypes;
 BlockIconEnd.args = {
   ...defaultArgs,
   icon: 'chevron-small-right-small',
-  'icon-flip': true,
   'icon-placement': iconPlacement.options[1],
   'text-size': textSize.options[0],
 };
@@ -264,7 +257,6 @@ BlockNegativeIconEnd.argTypes = defaultArgTypes;
 BlockNegativeIconEnd.args = {
   ...defaultArgs,
   icon: 'chevron-small-right-small',
-  'icon-flip': true,
   'icon-placement': iconPlacement.options[1],
   'text-size': textSize.options[0],
   negative: true,
@@ -301,6 +293,46 @@ InlineNegative.documentation = {
   title: 'Inline Negative',
 };
 
+const defaultArgTypesButton = {
+  disabled,
+  'event-id': eventId,
+  negative,
+  icon,
+  'icon-placement': iconPlacement,
+  'id-value': idValue,
+  text,
+  type,
+  variant,
+};
+
+const defaultArgsButton = {
+  disabled: false,
+  'event-id': 'Event ID for button click',
+  negative: false,
+  icon: '',
+  'icon-placement': iconPlacement.options[0],
+  'id-value': '',
+  text: 'Travelcards & tickets',
+  type: type.options[0],
+  variant: variant.options[1],
+};
+
+const InlineTemplateButton = (args, context) => (
+  <p style={paragraphStyle(context)}>
+    Lorem ipsum dolor sit amet. <sbb-link {...args}>{args.text}</sbb-link>
+  </p>
+);
+
+export const InlineButton = InlineTemplateButton.bind({});
+
+InlineButton.argTypes = defaultArgTypesButton;
+
+InlineButton.args = defaultArgsButton;
+
+InlineNegative.documentation = {
+  title: 'Inline Button',
+};
+
 export default {
   decorators: [
     (Story, context) => (
@@ -310,6 +342,9 @@ export default {
     ),
   ],
   parameters: {
+    actions: {
+      handles: [events.click],
+    },
     docs: {
       extractComponentDescription: () => readme,
     },